@@ -1,20 +1,18 @@
 {
   "name": "tr46",
-  "version": "3.0.0",
+  "version": "5.1.1",
   "engines": {
-    "node": ">=12"
+    "node": ">=18"
   },
   "description": "An implementation of the Unicode UTS #46: Unicode IDNA Compatibility Processing",
   "main": "index.js",
   "files": [
     "index.js",
-    "lib/mappingTable.json",
-    "lib/regexes.js",
-    "lib/statusMapping.js"
+    "lib/"
   ],
   "scripts": {
-    "test": "mocha",
-    "lint": "eslint .",
+    "test": "node --test",
+    "lint": "eslint",
     "pretest": "node scripts/getLatestTests.js",
     "prepublish": "node scripts/generateMappingTable.js && node scripts/generateRegexes.js"
   },
@@ -33,15 +31,14 @@
   ],
   "license": "MIT",
   "dependencies": {
-    "punycode": "^2.1.1"
+    "punycode": "^2.3.1"
   },
   "devDependencies": {
-    "@domenic/eslint-config": "^1.4.0",
-    "@unicode/unicode-14.0.0": "^1.2.1",
-    "eslint": "^7.32.0",
-    "minipass-fetch": "^1.4.1",
-    "mocha": "^9.1.1",
+    "@domenic/eslint-config": "^4.0.1",
+    "@unicode/unicode-16.0.0": "^1.6.5",
+    "eslint": "^9.22.0",
+    "globals": "^16.0.0",
     "regenerate": "^1.4.2"
   },
-  "unicodeVersion": "14.0.0"
+  "unicodeVersion": "16.0.0"
 }
