@@ -1,6 +1,6 @@
 {
   "name": "whatwg-url",
-  "version": "11.0.0",
+  "version": "14.2.0",
   "description": "An implementation of the WHATWG URL Standard's URL API and parsing machinery",
   "main": "index.js",
   "files": [
@@ -12,47 +12,42 @@
   "license": "MIT",
   "repository": "jsdom/whatwg-url",
   "dependencies": {
-    "tr46": "^3.0.0",
+    "tr46": "^5.1.0",
     "webidl-conversions": "^7.0.0"
   },
   "devDependencies": {
-    "@domenic/eslint-config": "^1.4.0",
+    "@domenic/eslint-config": "^4.0.1",
     "benchmark": "^2.1.4",
-    "browserify": "^17.0.0",
-    "domexception": "^4.0.0",
-    "eslint": "^7.32.0",
-    "got": "^11.8.2",
-    "jest": "^27.2.4",
-    "webidl2js": "^17.0.0"
+    "c8": "^10.1.3",
+    "esbuild": "^0.25.1",
+    "eslint": "^9.22.0",
+    "globals": "^16.0.0",
+    "webidl2js": "^18.0.0"
   },
   "engines": {
-    "node": ">=12"
+    "node": ">=18"
   },
   "scripts": {
-    "coverage": "jest --coverage",
-    "lint": "eslint .",
+    "coverage": "c8 node --test --experimental-test-coverage test/*.js",
+    "lint": "eslint",
     "prepare": "node scripts/transform.js",
     "pretest": "node scripts/get-latest-platform-tests.js && node scripts/transform.js",
-    "build-live-viewer": "browserify index.js --standalone whatwgURL > live-viewer/whatwg-url.js",
-    "test": "jest"
+    "build-live-viewer": "esbuild --bundle --format=esm --sourcemap --outfile=live-viewer/whatwg-url.mjs index.js",
+    "test": "node --test test/*.js",
+    "bench": "node scripts/benchmark.js"
   },
-  "jest": {
-    "collectCoverageFrom": [
-      "lib/**/*.js",
-      "!lib/utils.js"
+  "c8": {
+    "reporter": [
+      "text",
+      "html"
     ],
-    "coverageDirectory": "coverage",
-    "coverageReporters": [
-      "lcov",
-      "text-summary"
-    ],
-    "testEnvironment": "node",
-    "testMatch": [
-      "<rootDir>/test/**/*.js"
-    ],
-    "testPathIgnorePatterns": [
-      "^<rootDir>/test/testharness.js$",
-      "^<rootDir>/test/web-platform-tests/"
+    "exclude": [
+      "lib/Function.js",
+      "lib/URL.js",
+      "lib/URLSearchParams.js",
+      "lib/utils.js",
+      "scripts/",
+      "test/"
     ]
   }
 }
